requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4
openai>=1.12.0
python-dotenv>=1.0.1
neo4j>=5.14.0
python-dateutil>=2.8.2
